// Project/types/enums.ts
export enum DayOfTheWeekCode {
  Monday = 1,
  Tuesday = 2,
  Wednesday = 3,
  Thursday = 4,
  Friday = 5,
  Saturday = 6,
  Sunday = 7,
}

export enum TimeCode {
  Day = 1,
  Evening = 2,
  Night = 3,
}

export enum PlatformCode {
  PC = 'pc',
  Playstation = 'ps',
  Xbox = 'xb',
  Switch = 'ns',
}

export enum Pronoun {
  HE = 'he/him',
  SHE = 'she/her',
  THEY = 'they/them',
  HEY = 'he/they',
  SHEY = 'she/they',
}

export enum Flair {
  LGBTQ = 'LGBTQ+',
  STREAMER = 'Streamer',
  CREATOR = 'Content Creator',
  OTP = 'One-Trick Pony',
  JACK = 'Jack-of-All-Trades',
  COLLECTOR = 'Cosmetic Collector',
  NEWBIE = 'Newbie',
  VETERAN = 'Seasoned Veteran',
}

export enum ValorantRank {
  UNRANKED = 'Unranked',
  IRON = 'Iron',
  BRONZE = 'Bronze',
  SILVER = 'Silver',
  GOLD = 'Gold',
  PLATINUM = 'Platinum',
  DIAMOND = 'Diamond',
  ASCENDANT = 'Ascendant',
  IMMORTAL = 'Immortal',
  RADIANT = 'Radiant',
}

export enum ValorantRole {
  SENTINEL = 'Sentinel',
  DUELIST = 'Duelist',
  CONTROLLER = 'Controller',
  INITIATOR = 'Initiator',
}

export enum Region {
  BR = 'BR',
  EU = 'EU',
  KR = 'KR',
  LATAM = 'LATAM',
  NA = 'NA',
  AP = 'AP',
}

export enum IconForegroundKey {
  ACE = 'ACE',
  ALIEN = 'ALIEN',
  AXOLOTL = 'AXOLOTL',
  CAT = 'CAT',
  FISH = 'FISH',
  GLASSES = 'GLASSES',
  HERO = 'HERO',
  OCTOPUS = 'OCTOPUS',
  OWL = 'OWL',
  PIG = 'PIG',
  PIRATE = 'PIRATE',
  PUMPKIN = 'PUMPKIN',
  ROBOT = 'ROBOT',
  SKULL = 'SKULL',
  SPIDER = 'SPIDER',
  VAMPIRE = 'VAMPIRE',
  WOLF = 'WOLF',
  YINYANG = 'YINYANG',
}

export enum IconBackgroundKey {
  BLACK = '#041024',
  GREEN = '#06601b',
  BLUE = '#0c829f',
  PURPLE = '#9da3d8',
  PINK = '#C64D92',
  ORANGE = '#cd5c4b',
  YELLOW = '#dbA532',
}

export enum MatchingPreference {
  SIMILAR = 'Similar',
  DIFFERENT = 'Different',
  ANY = 'Any',
}

export enum CommunicationMethod {
  Text = -1,
  Voice = 1,
}

export enum HeatLevel {
  Chill = -1,
  Spicy = 1,
}

export enum DecisionMakingLevel {
  Strategic = 1,
  Aggressive = 2,
  Mixed = 3,
}

export enum TryHardLevel {
  Casual = -1,
  Competitive = 1,
}